login and register back, folder structure front

+++ FinalProject/server/controllers/filmController.js
@@ -1,7 +1,20 @@
 const marvelModel = require("../models/filmModels")
+var jwt = require('jsonwebtoken');
+require('dotenv').config()
 
 const getAllFilms = async(req, res) => {
+    const token = req.headers.authorization
+    if (!token) {
+        return res.status(401).json({ message: "Authorization token required" });
+    }
     try {
+        const decoded = jwt.verify(token, process.env.JWT_SECRET);
+
+        if (decoded.role !== "admin") {
+           return res.json({
+                message: "you are not an admin!"
+            })
+        }
         const response = await marvelModel.find({})
         res.json(response)
     } catch (error) {
@@ -12,8 +25,19 @@ const getAllFilms = async(req, res) => {
 }
 
 const getFilmsById = async(req, res) => {
+    const token = req.headers.authorization
     const {id} = req.params;
+    if (!token) {
+        return res.status(401).json({ message: "Authorization token required" });
+    }
     try {
+        const decoded = jwt.verify(token, process.env.JWT_SECRET);
+
+        if (decoded.role !== "admin") {
+           return res.json({
+                message: "you are not an admin!"
+            })
+        }
         const response = await marvelModel.findById(id)
         res.json(response)
     } catch (error) {
@@ -24,8 +48,19 @@ const getFilmsById = async(req, res) => {
 }
 
 const deleteFilm =  async(req, res) => {
+    const token = req.headers.authorization
     const {id} = req.params;
+    if (!token) {
+        return res.status(401).json({ message: "Authorization token required" });
+    }
     try {
+        const decoded = jwt.verify(token, process.env.JWT_SECRET);
+
+        if (decoded.role !== "admin") {
+           return res.json({
+                message: "you are not an admin!"
+            })
+        }
         const deletedResponse = await marvelModel.findByIdAndDelete(id)
         res.json(deletedResponse)
     } catch (error) {
@@ -36,8 +71,19 @@ const deleteFilm =  async(req, res) => {
 }
 
 const postFilm = async(req, res) => {
+    const token = req.headers.authorization
+    if (!token) {
+        return res.status(401).json({ message: "Authorization token required" });
+    }
     try {
-        const mrvl = marvelModel({...req.body})
+        const decoded = jwt.verify(token, process.env.JWT_SECRET);
+
+        if (decoded.role !== "admin") {
+           return res.json({
+                message: "you are not an admin!"
+            })
+        }
+        const mrvl =  new marvelModel({...req.body})
         await mrvl.save()
         res.json(mrvl)
     } catch (error) {
@@ -48,8 +94,19 @@ const postFilm = async(req, res) => {
 }
 
 const editFilm = async(req,res) => {
+    const token = req.headers.authorization
     const {id} = req.params;
+    if (!token) {
+        return res.status(401).json({ message: "Authorization token required" });
+    }
     try {
+        const decoded = jwt.verify(token, process.env.JWT_SECRET);
+
+        if (decoded.role !== "admin") {
+           return res.json({
+                message: "you are not an admin!"
+            })
+        }
         const updatedFilm = await marvelModel.findByIdAndUpdate(id,
             {...req.body},
             {new: true}
